feat: restructure code into folders and files
Add base_classes folder for base classes. Currently there is only one
base class for Flying Enemies.

Add views folder for different views. Currently have the start screen
and the pause screen.
